Milestone3/pyproject.toml
[project]
name = "milestone3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "groq>=0.37.1",
    "huggingface-hub>=0.23.0",
    "langchain>=0.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.0.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.3",
    "neo4j>=5.0.0",
    "networkx>=3.6.1",
    "openai>=2.11.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.52.1",
    "transformers>=4.40.0",
    "torch>=2.4.0,<2.6.0",
    "torchvision>=0.19.0,<0.21.0",
    "torchaudio>=2.4.0,<2.6.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }